Add passing test for addlist function
